requirements.txt
scikit-learn==1.8.0
nltk==3.9.4
Unidecode==1.4.0
rich==14.3.3
python-dotenv
openai==2.31.0